+++ TestPlatform.sln
@@ -157,8 +157,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "perf", "perf", "{0D4DF78D-7
 		scripts\perf\perfconfig.csv = scripts\perf\perfconfig.csv
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "V1UnitTestProject", "test\TestAssets\V1UnitTestProject\V1UnitTestProject.csproj", "{1F29087C-30F4-4733-8B8A-9226145E8431}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -769,18 +767,6 @@ Global
 		{826CD5AF-44FA-40F6-B731-3980CADED8C0}.Release|x64.Build.0 = Release|Any CPU
 		{826CD5AF-44FA-40F6-B731-3980CADED8C0}.Release|x86.ActiveCfg = Release|Any CPU
 		{826CD5AF-44FA-40F6-B731-3980CADED8C0}.Release|x86.Build.0 = Release|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Debug|x64.Build.0 = Debug|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Debug|x86.Build.0 = Debug|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Release|x64.ActiveCfg = Release|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Release|x64.Build.0 = Release|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Release|x86.ActiveCfg = Release|Any CPU
-		{1F29087C-30F4-4733-8B8A-9226145E8431}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
