+++ test/TestAssets/MstestV1UnitTestProject/MstestV1UnitTestProjectTestSettings.testsettings
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <TestSettings
   id="a43e2f05-dc45-4ae8-a2a7-b3738cf0a3a4"
-  name="V1UnitTestProjectTestSettings"
+  name="MstestV1UnitTestProjectTestSettings"
   enableDefaultDataCollectors="false"
   xmlns="http://microsoft.com/schemas/VisualStudio/TeamTest/2010">
   <Description><!--_locID_text="Description1"-->These are default test settings for a local test run.</Description>
