+++ test/TestAssets/MstestV1UnitTestProject/UnitTest1.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace V1UnitTestProject
+namespace MstestV1UnitTestProject
 {
     using System.Diagnostics;
     using System.Threading;
