+++ test/TestAssets/TestAssets.sln/TestAssets.sln
@@ -29,6 +29,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PerfTestProject", "..\PerfT
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleRunners", "..\ConsoleRunners\ConsoleRunners.csproj", "{FBDB7A61-C1DB-4B24-B568-DA96AAD0DB8A}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MstestV1UnitTestProject", "..\MstestV1UnitTestProject\MstestV1UnitTestProject.csproj", "{E28DD78A-E4C1-48C1-B027-88576F669C73}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -87,6 +89,10 @@ Global
 		{FBDB7A61-C1DB-4B24-B568-DA96AAD0DB8A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FBDB7A61-C1DB-4B24-B568-DA96AAD0DB8A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FBDB7A61-C1DB-4B24-B568-DA96AAD0DB8A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E28DD78A-E4C1-48C1-B027-88576F669C73}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E28DD78A-E4C1-48C1-B027-88576F669C73}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E28DD78A-E4C1-48C1-B027-88576F669C73}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E28DD78A-E4C1-48C1-B027-88576F669C73}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
